isolate openCV python changes
