+++ examples/open_cv_playground/config.json
@@ -1,0 +1,6 @@
+{
+    "packages": ["opencv-python", "numpy", "Pillow"],
+    "files": {
+        "/static/invent.min.tar.gz": "./*"
+    }
+}
